+++ rtc-20180111/alibabacloud_rtc20180111/models/_create_app_agent_template_request.py
@@ -16,11 +16,13 @@ class CreateAppAgentTemplateRequest(DaraModel):
         asr_config: main_models.CreateAppAgentTemplateRequestAsrConfig = None,
         back_channel_config: main_models.CreateAppAgentTemplateRequestBackChannelConfig = None,
         chat_mode: int = None,
+        enable_video_understanding: bool = None,
         greeting: str = None,
         interrupt_config: main_models.CreateAppAgentTemplateRequestInterruptConfig = None,
         interrupt_mode: int = None,
         llm_config: main_models.CreateAppAgentTemplateRequestLlmConfig = None,
         name: str = None,
+        prefer_video: int = None,
         tts_config: main_models.CreateAppAgentTemplateRequestTtsConfig = None,
         type: int = None,
     ):
@@ -31,12 +33,14 @@ class CreateAppAgentTemplateRequest(DaraModel):
         self.asr_config = asr_config
         self.back_channel_config = back_channel_config
         self.chat_mode = chat_mode
+        self.enable_video_understanding = enable_video_understanding
         self.greeting = greeting
         self.interrupt_config = interrupt_config
         self.interrupt_mode = interrupt_mode
         self.llm_config = llm_config
         # This parameter is required.
         self.name = name
+        self.prefer_video = prefer_video
         self.tts_config = tts_config
         self.type = type
 
@@ -79,6 +83,9 @@ class CreateAppAgentTemplateRequest(DaraModel):
         if self.chat_mode is not None:
             result['ChatMode'] = self.chat_mode
 
+        if self.enable_video_understanding is not None:
+            result['EnableVideoUnderstanding'] = self.enable_video_understanding
+
         if self.greeting is not None:
             result['Greeting'] = self.greeting
 
@@ -93,6 +100,9 @@ class CreateAppAgentTemplateRequest(DaraModel):
 
         if self.name is not None:
             result['Name'] = self.name
+
+        if self.prefer_video is not None:
+            result['PreferVideo'] = self.prefer_video
 
         if self.tts_config is not None:
             result['TtsConfig'] = self.tts_config.to_map()
@@ -126,6 +136,9 @@ class CreateAppAgentTemplateRequest(DaraModel):
         if m.get('ChatMode') is not None:
             self.chat_mode = m.get('ChatMode')
 
+        if m.get('EnableVideoUnderstanding') is not None:
+            self.enable_video_understanding = m.get('EnableVideoUnderstanding')
+
         if m.get('Greeting') is not None:
             self.greeting = m.get('Greeting')
 
@@ -142,6 +155,9 @@ class CreateAppAgentTemplateRequest(DaraModel):
 
         if m.get('Name') is not None:
             self.name = m.get('Name')
+
+        if m.get('PreferVideo') is not None:
+            self.prefer_video = m.get('PreferVideo')
 
         if m.get('TtsConfig') is not None:
             temp_model = main_models.CreateAppAgentTemplateRequestTtsConfig()