+++ rtc-20180111/alibabacloud_rtc20180111/models/_describe_app_agent_templates_response_body.py
@@ -76,12 +76,14 @@ class DescribeAppAgentTemplatesResponseBodyTemplates(DaraModel):
         back_channel_config: main_models.DescribeAppAgentTemplatesResponseBodyTemplatesBackChannelConfig = None,
         chat_mode: int = None,
         create_time: str = None,
+        enable_video_understanding: bool = None,
         greeting: str = None,
         id: str = None,
         interrupt_config: main_models.DescribeAppAgentTemplatesResponseBodyTemplatesInterruptConfig = None,
         interrupt_mode: int = None,
         llm_config: main_models.DescribeAppAgentTemplatesResponseBodyTemplatesLlmConfig = None,
         name: str = None,
+        prefer_video: int = None,
         tts_config: main_models.DescribeAppAgentTemplatesResponseBodyTemplatesTtsConfig = None,
         type: int = None,
     ):
@@ -91,12 +93,14 @@ class DescribeAppAgentTemplatesResponseBodyTemplates(DaraModel):
         self.back_channel_config = back_channel_config
         self.chat_mode = chat_mode
         self.create_time = create_time
+        self.enable_video_understanding = enable_video_understanding
         self.greeting = greeting
         self.id = id
         self.interrupt_config = interrupt_config
         self.interrupt_mode = interrupt_mode
         self.llm_config = llm_config
         self.name = name
+        self.prefer_video = prefer_video
         self.tts_config = tts_config
         self.type = type
 
@@ -139,6 +143,9 @@ class DescribeAppAgentTemplatesResponseBodyTemplates(DaraModel):
         if self.create_time is not None:
             result['CreateTime'] = self.create_time
 
+        if self.enable_video_understanding is not None:
+            result['EnableVideoUnderstanding'] = self.enable_video_understanding
+
         if self.greeting is not None:
             result['Greeting'] = self.greeting
 
@@ -156,6 +163,9 @@ class DescribeAppAgentTemplatesResponseBodyTemplates(DaraModel):
 
         if self.name is not None:
             result['Name'] = self.name
+
+        if self.prefer_video is not None:
+            result['PreferVideo'] = self.prefer_video
 
         if self.tts_config is not None:
             result['TtsConfig'] = self.tts_config.to_map()
@@ -189,6 +199,9 @@ class DescribeAppAgentTemplatesResponseBodyTemplates(DaraModel):
         if m.get('CreateTime') is not None:
             self.create_time = m.get('CreateTime')
 
+        if m.get('EnableVideoUnderstanding') is not None:
+            self.enable_video_understanding = m.get('EnableVideoUnderstanding')
+
         if m.get('Greeting') is not None:
             self.greeting = m.get('Greeting')
 
@@ -208,6 +221,9 @@ class DescribeAppAgentTemplatesResponseBodyTemplates(DaraModel):
 
         if m.get('Name') is not None:
             self.name = m.get('Name')
+
+        if m.get('PreferVideo') is not None:
+            self.prefer_video = m.get('PreferVideo')
 
         if m.get('TtsConfig') is not None:
             temp_model = main_models.DescribeAppAgentTemplatesResponseBodyTemplatesTtsConfig()